test(unit): converted the testing scaffolder tests to vitest
for #452
